Move Auth service to `service`
